pyproject.toml
[project]
name = "olt-chatbot"
version = "0.5.0"
description = "A chainlit chatbot for Olympiatoppen"
readme = "README.md"
requires-python = ">=3.12"
authors = [
    { name = "Martin Høy", email = "martin.hoy@dnv.com" },
    { name = "Luca Garré", email = "luca.garre@dnv.com" },
    { name = "Ingrid Eythorsdottir", email = "ingrid.eythorsdottir@olympiatoppen.no" },
]

dependencies = [
    "loguru >= 0.7.2",
    "python-dotenv >= 1.0.1",
    "pydantic == 2.10.1",           # Pydantic 2.10.2/3 breaks chainlit
    "langchain >= 0.3.7",
    "langchain-community >= 0.3.7",
    "langchain-openai >= 0.2.8",
    "langchain-chroma >= 0.1.4",
    "beautifulsoup4 >= 4.12.3",
    "rank-bm25 >= 0.2.2",
    "chainlit >= 1.3.2",
    "pymupdf >= 1.24.13",
    "pypdf2 >= 3.0.1",
    "pre-commit>=4.0.1",
]

[dependency-groups]
dev = [
    "pre-commit>=4.0.1",
    "types-beautifulsoup4 >= 4.12.0.20240229",
    "ipykernel >= 6.29.5",
]

# Define CLI commands
[project.scripts]
update-retrievers = "olt_chatbot.main:update_retrievers"
start-chainlit = "olt_chatbot.main:start_chainlit_app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"


[tool.ruff]
# Consider packages in the src directory as first-party code
src = ["src"]
# Run ruff on Notebooks as well
extend-include = ["*.ipynb"]

[tool.ruff.lint.per-file-ignores]
# Allow the use of assert in tests
"tests/*" = ["S101"]

[tool.ruff.lint]
select = [
    "F",   # Pyflakes
    "E",   # pycodestyle
    "W",   # pycodestyle
    "I",   # isort
    "ICN", # flake8-import-conventions
    "D",   # pydocstyle
    "N",   # pep8-naming
    "A",   # flake8-builtins
    "B",   # flake8-bugbear
    "S",   # flake8-bandit
    "SIM", # flake8-simplify
    "C4",  # flake8-comprehensions
    "DTZ", # flake8-datetimez
    "FA",  # flake8-future-annotations
    "ISC", # flake8-implicit-str-concat
    "PT",  # flake8-pytest-style
    "Q",   # flake8-quotes
    "RET", # flake8-return
    "PTH", # flake8-use-pathlib
    "UP",  # pyupgrade
    "RUF", # Ruff-specific rules
    "PD",  # pandas-vet
    # "FAST", # FastAPI
]
ignore = [
    "ISC001", # may cause conflicts when used with the formatter
]

[tool.ruff.lint.pydocstyle]
convention = "google"


# mypy configuration, taken from pydantic documentation
[tool.mypy]
plugins = ["pydantic.mypy"]

follow_imports = "silent"
warn_redundant_casts = true
warn_unused_ignores = true
disallow_any_generics = true
check_untyped_defs = true
no_implicit_reexport = true

disallow_untyped_defs = true

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true
